feat: manage HA intents
The previous logic was overly complex and prone to failure. This logic is configurable entirely from the skill side and defaults to on. In the case of Neon's Mark 2 image, where Home Assistant is shipped by default but perhaps not used, the skill setting `disable_intents` can be set to True in the image.